Performanc optimize cluster mode
Optimize cluster mode by utilizing the `cluster` module in the `index.js` file. If the file is being executed in the master mode, it will spawn child processes using `cluster.fork()`. On the other hand, if it is executed in child mode, it will act as a server and do nothing else. This change aims to improve the performance and scalability of the application.